Work on CSD3

+++ csd3/finemap.R
@@ -1,4 +1,4 @@
-# 17-9-2019 JHZ
+# 18-9-2019 JHZ
 
 zld <- function(z)
 {
@@ -26,20 +26,10 @@ pip.log10bf.plot <- function()
 
 options(digits=3, scipen=20, width=500)
 pr <- Sys.getenv("pr")
-
 z <- read.table(paste0(pr, ".z"), as.is=TRUE, header=TRUE)
 ld <- read.table(paste0(pr, ".ld.gz"),col.names=with(z,rsid))
 snp <- read.table(paste0(pr, ".snp"), as.is=TRUE, header=TRUE)
-topz <- subset(snp, abs(z)>=6.47)
-zldz <- zld(topz)
-topsnp <- head(snp, nrow(topz))
-zldsnp <- zld(topsnp)
-pip.log10bf.plot()
-load(paste0(pr,".rda"))
 snp <- within(snp, {rank <- 1:nrow(snp)})
-d <- merge(snpid_rsid,snp,by="rsid",all.y=TRUE)
-ord <-order(with(d,rank))
-snp <- d[ord,setdiff(names(d),c("chromosome","position","allele1","allele2","maf","beta","se","rank"))]
 config <- read.table(paste0(pr,".config"),as.is=TRUE,header=TRUE,nrows=31)
 if (file.exists(paste0(pr,".cred"))) cred <- read.table(paste0(pr,".cred"),as.is=TRUE,header=TRUE)
 
@@ -47,23 +37,41 @@ library(openxlsx)
 xlsx <- paste0(pr, "-finemap.xlsx")
 unlink(xlsx, recursive = FALSE, force = TRUE)
 wb <- createWorkbook(xlsx)
-addWorksheet(wb, "snp")
-writeDataTable(wb, "snp", snp)
-addWorksheet(wb, "pip.plot")
-insertImage(wb, "pip.plot", paste0(pr,".png"),height=12,width=8)
-addWorksheet(wb, "config")
-writeDataTable(wb, "config", config)
-if (exists("cred")) {
-  addWorksheet(wb, "cred")
-  writeDataTable(wb, "cred", cred)
-}
+# snp
+  load(paste0(pr,".rda"))
+  d <- merge(snpid_rsid,snp,by="rsid",all.y=TRUE)
+  ord <-order(with(d,rank))
+  name_snp <- d[ord,setdiff(names(d),c("chromosome","position","allele1","allele2","maf","beta","se","rank"))]
+  addWorksheet(wb, "snp")
+  writeDataTable(wb, "snp", name_snp)
+# pip.plot
+  pip.log10bf.plot()
+  addWorksheet(wb, "pip.plot")
+  insertImage(wb, "pip.plot", paste0(pr,".png"),height=12,width=8)
+# config
+  addWorksheet(wb, "config")
+  writeDataTable(wb, "config", config)
+# cred
+  if (exists("cred")) {
+    addWorksheet(wb, "cred")
+    writeDataTable(wb, "cred", cred)
+  }
 saveWorkbook(wb, file=xlsx, overwrite=TRUE)
 
-xlsx <- paste0(pr, ".zld.xlsx")
-unlink(xlsx, recursive = FALSE, force = TRUE)
-wb <- createWorkbook(xlsx)
-addWorksheet(wb, "topz")
-writeDataTable(wb, "topz", zldz)
-addWorksheet(wb, "topsnp")
-writeDataTable(wb, "topsnp", zldsnp)
-saveWorkbook(wb, file=xlsx, overwrite=TRUE)
+topz <- subset(snp, abs(z)>=6.47)
+if (nrow(topz) > 0)
+{
+# topz
+  zldz <- zld(topz)
+  topsnp <- head(snp, nrow(topz))
+  zldsnp <- zld(topsnp)
+  xlsx <- paste0(pr, ".zld.xlsx")
+  unlink(xlsx, recursive = FALSE, force = TRUE)
+  wb <- createWorkbook(xlsx)
+  addWorksheet(wb, "topz")
+  writeDataTable(wb, "topz", zldz)
+# topsnp
+  addWorksheet(wb, "topsnp")
+  writeDataTable(wb, "topsnp", zldsnp)
+  saveWorkbook(wb, file=xlsx, overwrite=TRUE)
+}
